ReplicationIT#shouldReplicateNewBranch: Create branch on local project
The test creates the branch directly on the target project, which
defeats the point of the replication test since the branch is not
actually getting replicated there.

Fix it to create the branch on the local project, which was the
original intention when the test was added in change I67aa8b4ae.

Change-Id: I517f3f49241207861e2f5158c08f6193a63d8916