Persist replication task when pushing all refs
Create a replication event associated with the replication of all
refs, so that the completion of the operation would find a corresponding
task and will not throw an exception.

Another benefit also is the ability to replay pending replication
events for a push of all refs upon plugin restarts and configuration
reloads.

Bug: Issue 11424
Change-Id: I63a82ff128c98e0ab7e58b76fa8f8548eb74e915