Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > It would be better to put this on stable-2.16 where the ReplicationIT was introduced.
> 
> Sure, but the same could be said for all the other removals of the references to EasyMock.
> This change isn't really a "bug-fix" so it could stay in master only IMHO.

It's more likely that we will add patches that touch ReplicationIT on stable-2.16 than on any of the other files that are getting migrated to mockito.  So putting this on stable-2.16 will reduce the possibility of conflicts in future.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>