Update patch set 1
Patch Set 1: Code-Review-1

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > It would be better to put this on stable-2.16 where the ReplicationIT was introduced.
> > 
> > Sure, but the same could be said for all the other removals of the references to EasyMock.
> > This change isn't really a "bug-fix" so it could stay in master only IMHO.
> 
> It's more likely that we will add patches that touch ReplicationIT on stable-2.16 than on any of the other files that are getting migrated to mockito.  So putting this on stable-2.16 will reduce the possibility of conflicts in future.

I agree we should fix it where this test was introduced.

As it is stated in the commit message, eiter it is an integration test or it is a mocked test, but it is never both. This test class is inherited from LightweightPluginDaemonTest and thus from AbstractDaemonTest and uis booting the whole GerritServer with replication plugin on top of it. So, there is no singl reason to add any mocks to full featured Gerrit server.

So that it was a bug from the beginning to depend on any Mock-Framework in Integration test, no matter, what mock framework is used.

We should check in core and in other plugins for this fundamental bug as well: mixing real integration tests with mocking parts.

Patch-set: 1
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1