Wrap calls to createProject in a utility method
The createProject method is removed in stable-3.0, and creating a test
project must be done with the ProjectOperations class. This means that
when a new test is added in stable-2.16 and uses createProject, there
is a conflict when merging it up to stable-3.0.

Wrap all the calls to createProject in a new utility method and call
that instead in the tests. This commit will conflict on the next merge
up to stable-3.0, but then there will be no more conflicts in future.

Change-Id: I271d0d55cd07c25eb4c6f8acc47d144b47600356