DestinationsCollection: synchronize shuttingDown flag
ShuttingDown flag is used to notify all other threads that
the collection of replication destinations is about to be stopped.

For instance, the auto-reload background thread should be informed
to not reload the configuration during shutdowns.
Having the flag as synchronized make sure that the flag is not
read and written concurrently by the shutdown and reload threads.

Change-Id: I6aa0a38d2c79a77befa5fde84da3377a1b0e7e93