Fix issue with dropping events on start
During the configuration reload ReplicationQueue dropped reference
updated events. Because replication scheduling is not created and
events are not preserved replication for these events never happened.

Solution is to add in memory queue which holds events until plugin
is up and running and then fire those events to make sure that all
replication happened.

Bug: Issue 11573
Change-Id: I3f9767c3879152acc8ff4ab73233abb278fb07a7