Clean up ReplicationQueue tests
ReplicationQueueIT integration test class was introduced in stable-2.15
branch to test fix for dropping events during the start. In stable-2.16
there is a ReplicationIT test class which contains tests for the same
part of the code. Merging both classes allows to avoid a lot of
boilerplate code and making tests easier to understand.

Bug: Issue 11573
Change-Id: I2ec71933f3657401932f1b3a1f371f0784a374a4