Fix firing pending "..all.." events on startup
The Destination.wouldPushRef() method is called on startup to see if a
Destination is configured to be pushed to for a specific ref, if it is
not configured to do so, then firing the pending update is skipped.
Since the magic "..all.." ref will never match the configuration in
replication.config, allways match it since if replication is configured
at all, then it should be matched.

Change-Id: I53bd527932e6aea9ddd465772925d601aa034bd3