Update patch set 3
Patch Set 3: Code-Review-1

(19 comments)

I like the change, but it tries to do way too much compared to what is mentioned in the commit message.

That is a symptom that it should be broken down in separate changes.
There are also implementation details leaking into the interface, which is exactly what this change is willing to avoid, isn't it?

Patch-set: 3
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1