Persist tasks before scheduling replication
Consolidate the replication firing paths and in doing so improve the
replicate-all path to persist events before scheduling them, thus
reducing the chance the task could be lost on server interruption. This
consolidation also has the side effect of adding replicate-all events to
the "beforeStartupEventsQueue" if they somehow happen to get fired
before replication is started.

Change-Id: Iaadbb89c4e5bc2f0eb751d26059868ba9cb7e9d7