Update patch set 3
Patch Set 3: Code-Review+1

I am wondering if the push "..all.." should be persisted or not as an event.
It is not clearly generated by a ref-update and it is typically done as a one-off admins step.

With regards to the push of all repos at startup, that is done regardless of the persisted event.

Could the fix be, instead, skipping the persistence of that push altogether? What do you think?

Patch-set: 3
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1