Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Good catch, this should be fairly easy to cover with a simple test, so that we make sure not to break it again in the future.

I don't believe it is possible to create a test that will catch this repeatedly. The stability does not change on command, and I don't know what changes it. It could be that it only changes on server startup, or when the moon turns blue...

Patch-set: 1
Reviewer: Gerrit User 1003883 <1003883@173816e5-2b9a-37c3-8a2e-48639d4f1153>