Decouple observable queue implementation from ReplicationQueue
Prior to this change, AutoReloadRunnable used to depend on
ReplicationQueue, which was used just to observe the status of the queue
during the run() phase of the reloading.

Decouple reloading of configuration from specific queue implementations
by depending only a way to observe the queue, rather than the queue
itself.

Incidentally, allow AutoReloadRunnable to be reused by implementations
that use a different queueing systems, or no queuing at all, such
as the pull-replication.

Change-Id: I589c5bf3fb49a53734ba3a991d08db33bc189d3e