Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I don't know how to review this because I don't know where this code will be used, is there a change in the other plugin that depends on this?

well, this is a fair question :)

replication plugin and pull-replication plugins both share the same replication.config file.

The rationale behind is that the same set of options, configurations and stanzas is broadly available and applicable to both, with the exception of the "direction" of the replication.

When a remote is configured with a "push" refSpec, then the target is a destination to replicate to (and it should be handled by the replication plugin).

When the remote configured with a "fetch" refSpec, then the target is a source to replicate from (and it should be handled by the pull-replication plugin).

The counter part of this change in the pull-replication plugin is this snippet of code[1], where remotes that have no "fetch" refSpec are ignored.

Hope this gives enough context for this to be reviewed.

https://gerrit.googlesource.com/plugins/pull-replication/+/c229fe435123df08afb92ed0beac977c60a6286f/src/main/java/com/googlesource/gerrit/plugins/replication/pull/SourcesCollection.java#171

Patch-set: 4
Reviewer: Gerrit User 1072905 <1072905@173816e5-2b9a-37c3-8a2e-48639d4f1153>