Allow to enforce replication of some projects on startup
A new configuration option replicateProjectsOnStartup has been added
to the ReplicationConfig. Projects specified in this option will be
always replicated, even if replicateOnStartup is set to false.

This is useful for a replica that is used solely for CI purposes.
In such a case it is particularly important that all the new changes
get replicated as soon as possible but it is way less important to
replicate anything else, except of All-Projects and All-Users.

This change allows enforcing replication of All-Projects and All-Users
to make sure that permissions and user accounts are replicated to
remotes on startup in case replication was not happening for some time.

Change-Id: I7275fee555d622fb21417a2df4087215b57b0dbd