Allow to enforce replication of some projects on startup
A new configuration option replicateProjectsOnStartup has been added
to the ReplicationConfig. Projects specified in this option will be
always replicated, even if replicateOnStartup is set to false.

A use case for this is enforcing replication of All-Projects
to make sure that permissions are replicated to remotes on startup
in case replication was not happening for some time.

Change-Id: I7275fee555d622fb21417a2df4087215b57b0dbd