Add REST interface for the replication start command
Implement a REST API that is the equivalent of the ssh
"start replication" command detailed in the existing
replication plugin documentation. To mimic the functionality
of the ssh command, keywords are used in the body rather than
add multiple REST endpoints. This feature is implemented in
the 3.1 branch as that is the target for a related internal
project.

Change-Id: I79a3ef0f78fe684150fb6b8915851af412eeaa40