Add GenericHttp implementation of the AdminApi
This is yet another way to perform remote admin operations
(creation/deletion of repositories). When replicating to a Gerrit slave,
we typically have an apache server running git-http-backend CGI script
to receive pushes from the master server. We cannot use the Gerrit slave
process to perform admin operations as Gerrit slave cannot perform any
write operations and doesn't even serve the REST API. Using the remote
SSH would be possible but would require setting up yet another
authentication path. However, when an apache server is setup, an
authentication channel is already setup and it could be handy to reuse
the same authentication for execution admin operations. One could easily
implement additional CGI script which performs project creation/deletion
and deploy it in the same apache server which runs the git-http-backend.

The proposed generic HTTP admin api will perform a PUT request on the
configured adminUrl to create a project and a DELETE request to delete a
project.

Change-Id: I23741576a58376cb114b17e029d7388c5f9ac54d