fixup! Extract destinations logic into a new class
The config file is never loaded by ReplicationFileBasedConfig and thus
is empty in its constructor. config.load() is later called by
DestinationsCollection.validateConfig(), but that doesn't affect the
values stored in ReplicationFileBasedConfig.

It seems DestinationsCollection.validateConfig() shouldn't be reading
config values or calling load() directly, but I'll address that in a
follow-up change.

Change-Id: I3f53237a1f0ac03948b72ae388722aed65716b4b