Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> Just a +1 to Saša's comments :-)

Seems like we agree on the approach to encapsulate the new config layout in a new
implementation of the ReplicationConfig interface: FanoutReplicationConfig.

As I already mentioned, currently there are many places in the replication plugin code
which depend on the ReplicationFileBasedConfig implementation class instead of depending
on the ReplicationConfig interface. We need to first get rid of that dependency and only
depend on the ReplicationConfig interface. This is a necessary precondition to implement
FanoutReplicationConfig and keep the rest of the plugin code stable with regards to any
changes in the storage format of replication configuration. Who is willing to work on that?

Patch-set: 2