Update patch set 1
Patch Set 1:

Why only CR+1 on this one?

FYI the change that is getting omitted is one that I added on stable-2.16; we use it at CollabNet but we're still on 2.15 and working on moving to 2.16.  We won't need be on 3.1 for some time, and I think I'll try to avoid needing it at all before we get there.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>