Fix slave going out of sync because of NFS delay
The problem with NFS is that it caches attributes, files and directories
information. A potential solution is to disable NFS caching but that
will degrade performance significantly.

The idea of this fix is to take into consideration NFS right before
building the delta that needs to be pushed to the slave.

Consider a potential file 'f' located on an NFS drive. Performing
f.exists() may returns false due to outdated cached information. Then
calling f.getParentFile().listFiles() helps to update NFS caches and
immediately makes f.exists() to return true.

A potential drawback is that this is doing "ls" operation every time
the method is being called which will result in minor addition to
replication delay time. However, with this fix there wont be missing
deltas during replication because some files were not updated in the
File System if NFS is used.

Change-Id: I0dca6612e8e6a51f9054d2ba97ce401b38db8cc2