Fix slave going out of sync because of NFS delay
The problem with NFS is that it caches attributes, files and directories
information, a potential solution is to disable NFS caching but that
will degrade performance significantly.

The idea of this fix is to take into consideration NFS right before
building the delta that needs to be pushed to the slave.

Consider a potential file 'f' located on an NFS drive. Performing
f.exists() will returns false due to outdated cached information.  Then
calling f.getParentFile().listFiles() helps to update NFS caches and
immediately makes f.exists() to return true.

A potential drawback is it is doing "ls" operation every time the method
is being called. However, each method execution will be more consistent
if NFS is used.

Change-Id: I0dca6612e8e6a51f9054d2ba97ce401b38db8cc2