Improve logging of why a project or ref is not replicated
Add more debug logs to indicate the cases where a project or
ref does not replicated.

For failure to read the project cache, log at warning level.

Also fix logging of the "scheduling replication ..." info log, which
was emitted before the check which could cause the replication to not
be scheduled. Move it to after the check, and add an additional debug
log of "NOT replicating ..." for the case where the check fails.

Change-Id: I13f5649cfe7e42fc672989efe8d77b47b0472023