PushOne: Do not attempt to send DELETE for refs/users/self
When pushing an "all refs" update to a remote, the list of refs
is obtained from the remote. If the project is All-Users and the
remote is a Gerrit server, then the ref list includes the magic
ref "refs/users/self".

Because the local refs are listed directly from the repo, they
don't include "refs/users/self". When replicating in "mirror" mode
this causes a DELETE command to be added to the push.

When the push is received on the remote, "refs/users/self" gets
translated back to the actual ref corresponding to the current user.
In the case where that user's ref is also being updated, the push is
rejected with:

  "cannot chain ref update DELETE after update UPDATE with different new ID"

Fix this by explicitly omitting "refs/users/self" from the DELETE
commands, in the same way as is already done for HEAD.

Change-Id: I52fff6d44692c2f7f54940f722e850828edb56ea