PushOne: Improve format of log of references to be pushed
The references to be pushed are logged by just passing the list
of RemoteRefUpdate instances to the logger. This results in the
List's default implementation of toString() being called, which
renders each object in a comma-separated list.

Each object is rendered by RemoteRefUpdate's toString() method
which includes some fields that are not relevant here, or omits
some information that might be useful. For example:

  [RemoteRefUpdate[remoteName=refs/meta/config, NOT_ATTEMPTED,
  (null)...a7038eb8827cfd29cb3fca335e882f4d5ed09b62, srcRef=refs/meta/config,
  forceUpdate, message=null]

- The 'remoteName' is the name on the destination. In this case it's
  the same as 'srcRef' but could be different if replication has been
  configured to push to a different refname. It is potentially confusing
  to have them not logged next to each other.
- '(null)...a7038eb8827cfd29cb3fca335e882f4d5ed09b62' shows the old
  and new IDs; this is in the standard git ref update format aside
  from '(null)' which is shown instead of zeros.
- 'forceUpdate' is only included when true; this is the same for
  the 'fastForward' field.
- the update has a flag indicating if it's a delete, but this is
  not included in the output at all.
- 'message' is always null for udates by replication so it's not
  necessary to include in the log.

Add a custom method to format the refs for logging, with the information
presented in a clearer way. Continue to log it on a single line, comma-
separated. This might not be ideal for human readers, but makes it easier
when processing the logs with grep or any other tool.

With this change, the logged information looks like:

  RemoteRefUpdate{refSpec=refs/meta/config:refs/meta/config,
  status=NOT_ATTEMPTED, id=(null)..a7038eb8827cfd29cb3fca335e882f4d5ed09b62
  force=yes, delete=no, ffwd=no}

Now the irrelevant information (message) is removed, missing information
(delete) is added, and the source and destination refs are presented in
a single "refSpec" field.

Change-Id: Idd153daf44cd79a0920ea0b72c64ea58d0926f59