PushOne: Improve logging of references being pushed
The references to be pushed are logged by just passing the list
of RemoteRefUpdate instances to the logger. This results in the
following behavior:

- The List's default implementation of toString is used, which
  renders each object in a comma-separated list
- Each object is rendered by RemoteRefUpdate's toString method
  which includes some fields that are not relevant here

This results in the refs being logged in a single line, which is
difficult to read when there are a lot of refs.

Add a custom method to format the refs for logging. Log each ref
on a single line, and only log the relevant pieces of info for each
ref. Invoke it lazily so that it's only done when the specific
logging level is enabled.

Change-Id: Idd153daf44cd79a0920ea0b72c64ea58d0926f59