Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This was previously done in [1].  Cherry-pick that one?
> 
> [1] https://gerrit-review.googlesource.com/c/plugins/replication/+/206729

Good point. I considered to cherry-pick Dave's change, but decided against it, because I assume that the whole series of changes wouldn't be submitted. It is only created to blindly try to fix Eclipse Foundation Gerrit site upgrade. Probably the batch mode operation deactivation in MySQL driver is not the root cause.

Patch-set: 1