Clarify the limitations of gerrit+ssh in replication.config
The gerrit+ssh option, introduced years ago when Lucene was not
used in Gerrit, is currently misunderstood by the users because
it is presented as a way to perform Gerrit master to master
replication.

Whilst it was actually possible in the past, it isn't supported
anymore because of the need of more machinery around it.

Clarify the limitations in the documentation and warn on the
limited applicability.

Bug: Issue 12616
Change-Id: I992385b9da03b0c3c2b938817687ba50b2dbde58