Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> I wonder if it's better to just remove the gerrit+ssh feature, or at least remove the documentation so it's no longer an officially supported feature?

Yeah, we can justify with the statement "has a very limited set of use-cases and thus isn't generally supported anymore".

Anyway, we are fully missing in the replication plugin documentation the high-level use-case of the setup: too many people make the mistake of pointing to Gerrit ports instead of a git-daemon. But that's a different concern, I'll address it in a separate change.

Patch-set: 1