Remove support for gerrit+{ssh,http} protocols
The gerrit+ssh and gerrit+http replication protocols are
not fully working because of the limitations of pushing to another
Gerrit master:

- refs/changes/*/meta cannot be pushed
- refs/meta/* for All-Users cannot be pushed
- refs/users/* for All-Users cannot be pushed

Also, the Gerrit master to master replication without the use
of a git-daemon as replication endpoint has never been fully
working end-to-end in recent releases (v2.16 or later).

Remove the feature of creating/removing repositories by pointing
the replication plugin directly to a Gerrit master SSH or HTTP
ports.

It would still be possible for a Gerrit admin to point a replication
end to a Gerrit master port, via standard Git/HTTP or Git/SSH,
bearing in mind that he will need to take further actions to
make that use-case working. It is beyond the scope of this change
to allow or forbid the use-case.

Change-Id: I961effe018239c9b3f413f674f14c0b47751774e