Remove support for gerrit+{ssh,http} protocols
The gerrit+ssh and gerrit+http replication protocols are
not fully working because of the limitations of pushing to another
Gerrit master:

- refs/changes/*/meta cannot be pushed
- refs/meta/* for All-Users cannot be pushed
- refs/users/* for All-Users cannot be pushed

Also, the Gerrit master to master replication without the use
of a git-daemon as replication endpoint has never been fully
working end-to-end without the support of extra logic for
changes reindexing and cache invalidation.

Remove the feature so that users in the future would not
fall into using it when it would not be fully working anyway.

Change-Id: I961effe018239c9b3f413f674f14c0b47751774e