Remove documentation for gerrit+{ssh,http} protocols
The gerrit+ssh and gerrit+http replication protocols are
not fully working because of the limitations of pushing to another
Gerrit master:

- refs/changes/*/meta cannot be pushed
- refs/meta/* for All-Users cannot be pushed
- refs/users/* for All-Users cannot be pushed

Also, the Gerrit master to master replication without the use
of a git-daemon as replication endpoint has never been fully
working end-to-end in recent releases (v2.16 or later).

The gerrit+ssh and gerrit+http are COLLABNET-specific options
that can work only with their fork of Gerrit, hence
tag the two APIs with UsedAt(UserAt.Project.COLLABNET)
and do not mention their use in the official Gerrit docs as
it would be misleading for the user.

Change-Id: I961effe018239c9b3f413f674f14c0b47751774e