Improve error messages related to other node actions
In a multi-master scenario with shared storage, it is common for
operations on one node to "interfere" with filesystem operations on
another node without causing a malfunction. Improve the logging in these
cases with likely explanation of the filesystem error, and do not
consider these specific filesystem errors operational errors. To achieve
these better error messages, the task storage "recover" part of the
"reset" API was split out into its own API, this has the added benefit
of making recovery more explicit.

Change-Id: Idd7149abf65ed66e1c05c59cb1afee17f7cb8622