Fix potential event loss during external collisions
Previously when a URI was being replicated to by a node in a shared
storage multi-master setup without a replication distributor enabled, a
new push could be lost. This loss would occur if another node attempted
to push to the same URI for a different newer push generated locally. In
this case, the second node would be denied the lock on the URI and
behave as if the push were already completed. As this push is for a
different newer event, the push would be dropped and lost. This
occurred because the code treated externally denied events, which could be created either
by external completions or external in-flight operations, the same. Now, in the
in case of externally in-flight pushes, treat them similar to local in-flight,
collisions and reschedule them.

Change-Id: I44e999e56c02cdca0ddc333b561ad8b2d861c479