Fix dealing with external collisions and completions
Fix potential event loss during external collisions, and prevent start
--wait from failing on external completions.  Previously externally
denied events, which could be created either by external completion or
external in-flight operations, were treated the same. Split these two
situations as the in-flight operation may still have new waiting tasks
that need to be rescheduled. Also be sure to update the statuses
correctly on completions so that they don't fall back to assumed
failures in the cleanup code.

Change-Id: I44e999e56c02cdca0ddc333b561ad8b2d861c479