Fix potential event loss during external collisions
Previously when a URI was being replicated to by a node in a shared
storage multi-master setup without a replication distributor enabled, a
new push could be lost. This lost would occur if another node attempted
to push to the same URI for a different newer push generated locally. In
this casei, the second node would be denied the lock on the URI and
behave as if the push were already completed. As this push is for a
different newer event, the push would be dropped and lost.  This
occurred because externally denied events, which could be created either
by external completion or external in-flight operations, were treated
the same. Split these two situations as the in-flight operation may
still have new waiting tasks that need to be rescheduled. Now, in the
case of an externally in-flight operation, treat the push as in-flight,
and reschedule it.

Change-Id: I44e999e56c02cdca0ddc333b561ad8b2d861c479