Update patch set 5
Patch Set 5:

Looks like this change broke the replication_log: the replication-task-id is not any more
present in the replication_log which makes it impossible to track and troubleshoot replication
tasks.

Before this change the entries in the replication_log contained task-id and we could simply
grep all entries by the task-id d45aeddf:

  [2020-04-19 00:05:04,921] [] scheduled my-project:refs/heads/master => [d45aeddf] push ... to run after 5s
  [2020-04-19 00:05:09,921] [d45aeddf] Rescheduling replication to ... to avoid collision with the in-flight push [2fc97d12].
  [2020-04-19 00:06:04,925] [d45aeddf] Replication to ... started...
  [2020-04-19 00:06:04,969] [d45aeddf] Push to ... references: 
  [RemoteRefUpdate[remoteName=refs/heads/master, NOT_ATTEMPTED, ... ], ...
  [2020-04-19 00:10:04,099] [d45aeddf] Replication to ... completed in 239174ms, 60100ms delay, 0 retries

Now, the task-id is only reported when the task is created and is not present in any subsequent
log entries:

  [2020-04-19 00:05:04,921] [] scheduled my-project:refs/heads/master => [d45aeddf] push ... to run after 5s
  [2020-04-19 00:05:09,921] [] Rescheduling replication to ... to avoid collision with the in-flight push [2fc97d12].
  [2020-04-19 00:06:04,925] [] Replication to ... started...
  [2020-04-19 00:06:04,969] [] Push to ... references: [RemoteRefUpdate[remoteName=refs/heads/master, NOT_ATTEMPTED, ... ], ...
  [2020-04-19 00:10:04,099] [] Replication to ... completed in 239174ms, 60100ms delay, 0 retries


Any idea how to fix this? Otherwise, we should revert this change.

Patch-set: 5
CC: Gerrit User 1003873 <1003873@173816e5-2b9a-37c3-8a2e-48639d4f1153>