Improve listing error messages related to other node actions
In a multi-master scenario with shared storage, it is common for
operations on one node to "interfere" with task listing operations on
another node without causing a malfunction. Improve the logging in these
cases with likely explanation of the listing error, and do not consider
these specific filesystem errors operational errors.

Change-Id: Ia2ad431c20142ff0ce23dbace34aec837e3d8540