Prevent persistent task listing interruptions on IOExceptions
When iterating over the list of persisted tasks, it is possible for an
IOException to occur while reading a specific task. Prevent this
exception from breaking out of the iteration by catching and logging the
exception inside the loop instead of outside of it.

Also improve the logging by differentiating between failures that are
severe versus potentially related to other node actions since in a
multi-master scenario with shared storage, it is common for operations
on one node to "interfere" with task listing operations on another node
without causing a malfunction.  Specifically, improve the exception
handling so that the logging in these latter cases have a likely
explanation of the listing error, and do not consider these specific
filesystem errors operational errors.

Change-Id: Ia2ad431c20142ff0ce23dbace34aec837e3d8540