ReplicationQueue: Revert back from Flogger to slf4j logging
Slf4j's Mapped Diagnostic Context (MDC) is used to add the replication Id
into the logs and this was not adapted in the migration to Flogger,
resulting in the replication Id being omitted from the logs.

Flogger does not support MDC in the same way as slf4j, but provides a log
context which could be used. This would however change the format of the log
output. An alternative would be to manually update all the log statements
to include the replication Id. The downside of this approach is that we
need to remember to manually add it in any new logs that get added.

Instead of attempting either of those, just revert back to using slf4j.

This reverts commit 425f4f20218986615593da1a9319466743ee12eb.

Note that this is not a pure revert of 425f4f2; following that commit
several new logs were added. This reverts the commit and adapts the new
logs to also use slf4j.

Bug: Issue 12678
Change-Id: I3002afdac7c31786de5b611b32fa3f4c1a80e657