Convert PushResultProcessing to an interface
Instead of being an abstract class with methods that must be explicitly
overridden (even if they are no-op), convert it to an interface with
default empty methods. Also introduce a NO_OP instance. This allows to
have implementations that do nothing without having to explicitly add
all the boilerplate overrides.

Change-Id: I8a3cf6fc66f9a3857e5fa4949ca99c2c473702d6