Don't lose state when there's a pending push to the same ref
Previously if there was already a pending push (not an in-flight push)
to the same endpoint, the start for the push would be dropped when
adding the push to the Destination. This meant that a replication start
--wait command would never complete when one of its pushes was pending
since its state would never receive the completion notification for that
push.

Fix this by always adding the state and the ref in the Destination class
and preventing the duplicate "addRef" log message in the PushOne class.

Bug: Issue 12731
Change-Id: I33e6af2709bc38aac791a2f60fd896492167bbf5