Fix start --wait to track in-flight collisions and to not fail
Previously when an in-flight collision occurred on a push that the
'start --wait' command  was waiting for, the command would terminate
with an error (and if it weren't for an NPE it would stop waiting
prematurely). This was happening because
PushOne is not tracking collisions resulting in its cleanup code
assuming there was a failure. Fix this by tracking the collision in
PushOne and skipping the cleanup in this case.

Ironically, the NPE in change I107d964a33349aaa5d6ae9aca68aab2889689155
was preventing the start command from receiving the premature
termination on the collision, but it was not preventing the error. Thus
this fix should get merged before fixing the NPE to avoid a worse
scenario.

Change-Id: Ibf1ca624739205f6f92e097213828faf29de9300