Allow extensibility of replication configuration parsing
Configuration parsing during the configuration auto reload is always
returning collection of DestinationConfiguration objects regardless of
the injected ConfigParser in the module. This situation is causing issue
for plugins that extends the replication plugin(e.g. pull-replication)
because they are expecting different remote configuration class(e.g.
SourceCollection instead of DestinationCollection). Solution is to
create ConfigParser interface which can be bind to an implementation
in a module.

Bug: Issue 12720
Change-Id: Iaec22e082f49eaac56b5555074bb80c269879b68