Fix extensibility of replication configuration parsing
Fix an issue where configuration parsing during the configuration
auto reload is always returning collection of DestinationConfiguration
objects regardless of the injected ConfigParser in the module.

This situation is causing a regression for plugins that extend
the replication plugin (e.g. pull-replication) because they are
expecting different remote configuration class (e.g.SourceCollection
instead of DestinationCollection).

Restore the pluggability of the configuration parsing by creating
a ConfigParser interface which can be bound to an implementation
in the plugin's Guice Module.

Bug: Issue 12720
Change-Id: Iaec22e082f49eaac56b5555074bb80c269879b68