Don't wait for pending events to process on startup
Previously, on large Gerrit installations with many projects and/or many
replication destinations, the replication plugin could take very long
periods of time to startup. This was particularly a problem if the
pending(persisted) event count was large as they all were rescheduled
before the plugin finished initializing. Change this behavior so that
startup merely begins the process of scheduling the pending events, but
does not wait for them to complete.

To achieve this, the processing of pending events is moved to the
defaultQueue, and this is done in a non invasive way so that this
processing should not interfere with other important tasks in the queue
for longer than it takes to process a single entry. This is achieved by
using a new ChainedScheduler class which breaks the workload up into
small pieces with their own task each, and only places a new task in the
queue each time a task begins execution. This creates a natural load
based throttling which can execute fast when there is no load in the
executor, but mostly gets out of the way when load is high.

Change-Id: Icddf88b85dec9078bb4a7923b9211614989978ff